services/auth-service/requirements.txt
# 认证服务依赖包

# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 认证和安全
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# 数据验证
pydantic[email]==2.5.0

# 邮件服务
sendgrid==6.10.0
aiosmtplib==3.0.1
jinja2==3.1.2

# Redis（会话和缓存）
redis==5.0.1
aioredis==2.0.1

# 工具库
python-dotenv==1.0.0
click==8.1.7

# 开发和测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0